requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
WTForms==3.0.1
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
gunicorn==21.2.0
markdown
email_validator